BUG: assignment to a variable from a user function segfaulted      check that the index returned is higher than the list->count, instead of equal to one more than the max
variables.c: split the increment and assignment to two different lines so I don't confuse myself in the future
operator.c/h: changes char const *const list[] to char const *const *const list, no real reason other than style
	      add -6 error for the failed malloc

+++ multi.c
@@ -556,7 +556,7 @@ struct matrix *extractValue(char **inp, int var_ind, struct vari *var, err_ret *
 err_ret chk_var(const char *buffer, char *inp, uint16_t *iter, struct vari *var, struct stack *num, struct stack *ch) {
 	err_ret error = 0;
 
-	uint16_t len = strlen(buffer);
+	int len = strlen(buffer);
 
 	char *nameBuffer = malloc(sizeof(*nameBuffer) * (len + 1));
 	__MALLOC_CHECK(nameBuffer, error);
